fix(cleanup): inactivate all deployment related to PR environment
Ensure cleanup of pull request related deployments by listing based on environment

Previously, listing GitHub deployments by git ref only worked with the ref used to create the deployment.
However, as the pull request ref can be a branch or a sha, there was a risk of failing to clean up some deployments.
To address this issue, we have updated the deployment listing to be based on the deployment's environment.
This ensures that we don't overlook any deployment related to the pull request, and can properly destroy or inactivate them as needed.